Fix conversion issue on maxRap after changing backing data to metric
